♻️ refactor: split channel/cli/chat.go into focused files
- chat_view.go: View(), resize(), renderMarkdown(), refreshViewport()
- chat_input.go: handleInput(), waitNextChunk(), updateCompletions()
- chat_picker.go: handlePickingKey(), currentModelIndex()
- chat.go: types, chatModel struct, newChatModel, Init(), Update()
