+++ channel/cli/chat.go
@@ -10,9 +10,6 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/glamour"
-	"github.com/charmbracelet/glamour/styles"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/muesli/termenv"
 
 	"github.com/vaayne/anna/agent"
 	"github.com/vaayne/anna/agent/runner"
@@ -353,349 +350,4 @@ func (m chatModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	cmds = append(cmds, cmd)
 
 	return m, tea.Batch(cmds...)
-}
-
-// updateCompletions shows/hides the command completion popup based on textarea content.
-func (m *chatModel) updateCompletions() {
-	val := m.textarea.Value()
-	if strings.HasPrefix(val, "/") {
-		matches := filterCommands(val)
-		wasCompleting := m.completing
-		m.completing = len(matches) > 0
-		m.completions = matches
-		if m.completeCursor >= len(matches) {
-			m.completeCursor = 0
-		}
-		if m.completing != wasCompleting {
-			m.resize()
-		}
-	} else if m.completing {
-		m.completing = false
-		m.completions = nil
-		m.completeCursor = 0
-		m.resize()
-	}
-}
-
-// horizontal padding on each side of the content area
-const padX = 1
-
-func (m *chatModel) resize() {
-	// Layout height budget (count actual rendered lines):
-	//   header (title):             1 line
-	//   blank line (\n\n):          1 line
-	//   viewport:                   vpHeight lines
-	//   input separator top:        1 line
-	//   prompt + textarea:          ta.Height lines
-	//   input separator bottom:     1 line
-	//   help bar:                   1 line
-	//   completion popup:           variable
-	const chrome = 1 + 1 + 1 + 1 + 1 + 1 // header + blank + sep top + sep bottom + newline + help
-
-	completionHeight := 0
-	if m.completing && len(m.completions) > 0 {
-		completionHeight = len(m.completions)
-	}
-
-	vpHeight := m.height - chrome - m.textarea.Height() - completionHeight
-	if vpHeight < 1 {
-		vpHeight = 1
-	}
-
-	innerWidth := m.width - padX*2
-
-	if !m.ready {
-		m.viewport = viewport.New(innerWidth, vpHeight)
-		m.viewport.SetContent(m.history.String())
-		m.ready = true
-	} else {
-		m.viewport.Width = innerWidth
-		m.viewport.Height = vpHeight
-	}
-
-	m.textarea.SetWidth(innerWidth - 2) // subtract prompt "> " width
-
-	// Recreate markdown renderer with no document margin for flush-left alignment
-	style := styles.DarkStyleConfig
-	if !termenv.HasDarkBackground() {
-		style = styles.LightStyleConfig
-	}
-	style.Document.Margin = uintPtr(0)
-	style.Document.BlockPrefix = ""
-	style.Document.BlockSuffix = ""
-	m.mdRenderer, _ = glamour.NewTermRenderer(
-		glamour.WithStyles(style),
-		glamour.WithWordWrap(innerWidth),
-	)
-}
-
-func (m *chatModel) handleInput(input string) tea.Cmd {
-	switch input {
-	case "/quit", "/exit":
-		return tea.Quit
-	case "/new":
-		info, err := m.pool.RotateSession(cliChannel)
-		if err != nil {
-			m.history.WriteString(errorStyle.Render("error: "+err.Error()) + "\n\n")
-		} else {
-			m.sessionID = info.ID
-			m.history.Reset()
-			m.historyPrefix = ""
-			m.history.WriteString(systemStyle.Render("[new session started]") + "\n\n")
-		}
-		m.viewport.SetContent(m.history.String())
-		m.viewport.GotoBottom()
-		return nil
-	case "/compact":
-		m.streaming = true
-		m.status = "Compacting session..."
-		m.textarea.Blur()
-		ctx := m.ctx
-		sessionID := m.sessionID
-		return func() tea.Msg {
-			summary, err := m.pool.CompactSession(ctx, sessionID)
-			return compactDoneMsg{summary: summary, err: err}
-		}
-	case "/model":
-		m.models = toModelOptions(m.listModels())
-		if len(m.models) == 0 {
-			m.history.WriteString(systemStyle.Render("[no models configured]") + "\n\n")
-			m.viewport.SetContent(m.history.String())
-			m.viewport.GotoBottom()
-			return nil
-		}
-		m.picking = true
-		m.modelFilter = ""
-		m.filteredModels = m.models
-		m.modelCursor = m.currentModelIndex()
-		m.textarea.Blur()
-		m.viewport.SetContent(renderModelPicker(m.filteredModels, m.modelCursor, m.provider, m.model, m.modelFilter))
-		m.viewport.GotoTop()
-		return nil
-	}
-
-	m.history.WriteString(userStyle.Render("You") + "\n" + userBorderStyle.Render(chatTextStyle.Render(input)) + "\n\n")
-	m.history.WriteString(agentStyle.Render("Anna") + "\n")
-	m.historyPrefix = m.history.String()
-	m.currentRaw.Reset()
-	m.viewport.SetContent(m.historyPrefix)
-	m.viewport.GotoBottom()
-
-	m.streaming = true
-	m.status = "Thinking..."
-	m.textarea.Blur()
-
-	ctx := m.ctx
-	sessionID := m.sessionID
-	return func() tea.Msg {
-		stream := m.pool.Chat(ctx, sessionID, input)
-		return streamStartMsg{stream: stream}
-	}
-}
-
-// waitNextChunk returns a Cmd that reads the next event from the stream channel.
-func waitNextChunk(stream <-chan runner.Event) tea.Cmd {
-	if stream == nil {
-		return nil
-	}
-	return func() tea.Msg {
-		evt, ok := <-stream
-		if !ok {
-			return streamDoneMsg{}
-		}
-		if evt.Err != nil {
-			return streamErrMsg{evt.Err}
-		}
-		if evt.ToolUse != nil {
-			return streamToolMsg{
-				tool:   evt.ToolUse.Tool,
-				status: evt.ToolUse.Status,
-				input:  evt.ToolUse.Input,
-				detail: evt.ToolUse.Detail,
-			}
-		}
-		return streamChunkMsg(evt.Text)
-	}
-}
-
-// renderMarkdown renders raw markdown text using glamour, falling back to raw text on error.
-func (m *chatModel) renderMarkdown(raw string) string {
-	if m.mdRenderer == nil || raw == "" {
-		return raw
-	}
-	rendered, err := m.mdRenderer.Render(raw)
-	if err != nil {
-		return raw
-	}
-	return strings.TrimRight(rendered, "\n")
-}
-
-// refreshViewport rebuilds viewport content from historyPrefix + rendered current response.
-func (m *chatModel) refreshViewport() {
-	rendered := m.renderMarkdown(m.currentRaw.String())
-	content := m.historyPrefix + agentBorderStyle.Render(rendered)
-	m.viewport.SetContent(content)
-	m.viewport.GotoBottom()
-}
-
-func uintPtr(v uint) *uint { return &v }
-
-// padLines prepends padding to each line of text.
-func padLines(text, pad string) string {
-	lines := strings.Split(text, "\n")
-	for i, line := range lines {
-		lines[i] = pad + line
-	}
-	return strings.Join(lines, "\n")
-}
-
-func (m chatModel) handlePickingKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.Type {
-	case tea.KeyCtrlC:
-		return m, tea.Quit
-
-	case tea.KeyEsc:
-		m.picking = false
-		m.viewport.SetContent(m.history.String())
-		m.viewport.GotoBottom()
-		m.textarea.Focus()
-		return m, nil
-
-	case tea.KeyUp:
-		if m.modelCursor > 0 {
-			m.modelCursor--
-		}
-		m.viewport.SetContent(renderModelPicker(m.filteredModels, m.modelCursor, m.provider, m.model, m.modelFilter))
-		return m, nil
-
-	case tea.KeyDown:
-		if m.modelCursor < len(m.filteredModels)-1 {
-			m.modelCursor++
-		}
-		m.viewport.SetContent(renderModelPicker(m.filteredModels, m.modelCursor, m.provider, m.model, m.modelFilter))
-		return m, nil
-
-	case tea.KeyEnter:
-		if len(m.filteredModels) == 0 {
-			return m, nil
-		}
-		selected := m.filteredModels[m.modelCursor]
-		m.picking = false
-
-		if selected.provider == m.provider && selected.model == m.model {
-			m.viewport.SetContent(m.history.String())
-			m.viewport.GotoBottom()
-			m.textarea.Focus()
-			return m, nil
-		}
-
-		if m.switchModel != nil {
-			if err := m.switchModel(selected.provider, selected.model); err != nil {
-				m.history.WriteString(errorStyle.Render("error switching model: "+err.Error()) + "\n\n")
-				m.viewport.SetContent(m.history.String())
-				m.viewport.GotoBottom()
-				m.textarea.Focus()
-				return m, nil
-			}
-		}
-
-		info, err := m.pool.RotateSession(cliChannel)
-		if err != nil {
-			m.history.WriteString(errorStyle.Render("error creating session: "+err.Error()) + "\n\n")
-		} else {
-			m.sessionID = info.ID
-		}
-
-		m.provider = selected.provider
-		m.model = selected.model
-		m.history.WriteString(systemStyle.Render(fmt.Sprintf("[switched to %s/%s]", m.provider, m.model)) + "\n\n")
-		m.viewport.SetContent(m.history.String())
-		m.viewport.GotoBottom()
-		m.textarea.Focus()
-		return m, nil
-
-	case tea.KeyBackspace:
-		if len(m.modelFilter) > 0 {
-			m.modelFilter = m.modelFilter[:len(m.modelFilter)-1]
-			m.filteredModels = filterModels(m.models, m.modelFilter)
-			m.modelCursor = 0
-			m.viewport.SetContent(renderModelPicker(m.filteredModels, m.modelCursor, m.provider, m.model, m.modelFilter))
-			m.viewport.GotoTop()
-		}
-		return m, nil
-
-	case tea.KeyRunes:
-		m.modelFilter += string(msg.Runes)
-		m.filteredModels = filterModels(m.models, m.modelFilter)
-		m.modelCursor = 0
-		m.viewport.SetContent(renderModelPicker(m.filteredModels, m.modelCursor, m.provider, m.model, m.modelFilter))
-		m.viewport.GotoTop()
-		return m, nil
-	}
-
-	return m, nil
-}
-
-func (m chatModel) currentModelIndex() int {
-	for i, opt := range m.filteredModels {
-		if opt.provider == m.provider && opt.model == m.model {
-			return i
-		}
-	}
-	return 0
-}
-
-func (m chatModel) View() string {
-	if !m.ready {
-		return "Initializing..."
-	}
-
-	pad := strings.Repeat(" ", padX)
-
-	// Title bar: "Anna" left, "provider/model" right
-	title := titleStyle.Render("Anna")
-	modelInfo := modelInfoStyle.Render(m.provider + "/" + m.model)
-	titleGap := m.width - padX*2 - lipgloss.Width(title) - lipgloss.Width(modelInfo)
-	if titleGap < 0 {
-		titleGap = 0
-	}
-	header := pad + title + strings.Repeat(" ", titleGap) + modelInfo + pad
-
-	// Input area — two thin separator lines with > prompt
-	sepLine := pad + inputSeparator.Render(strings.Repeat("─", m.width-padX*2))
-	prompt := inputPromptStyle.Render(">")
-	input := sepLine + "\n" + pad + prompt + " " + m.textarea.View() + "\n" + sepLine
-
-	// Completion popup
-	completionView := ""
-	if m.completing && len(m.completions) > 0 {
-		completionView = "\n" + renderCompletions(m.completions, m.completeCursor)
-	}
-
-	// Help bar below input
-	var helpText string
-	switch {
-	case m.picking:
-		helpText = helpStyle.Render("↑↓ · enter · esc")
-	case m.completing:
-		helpText = helpStyle.Render("↑↓ · tab · enter · esc")
-	default:
-		helpText = helpAccentStyle.Render("/new") + helpStyle.Render(" · ") +
-			helpAccentStyle.Render("/model") + helpStyle.Render(" · ") +
-			helpAccentStyle.Render("/quit")
-	}
-	status := ""
-	if m.status != "" {
-		status = statusStyle.Render(m.status)
-	}
-	helpGap := m.width - padX*2 - lipgloss.Width(helpText) - lipgloss.Width(status)
-	if helpGap < 0 {
-		helpGap = 0
-	}
-	helpBar := pad + helpText + strings.Repeat(" ", helpGap) + status
-
-	// Viewport with padding
-	vpView := padLines(m.viewport.View(), pad)
-
-	return header + "\n\n" + vpView + "\n" + input + completionView + "\n" + helpBar
 }
